requirements.txt
Click==7.0
enum34==1.1.6
Flask==1.0.2
futures==3.2.0
imageio==2.4.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.0
Pillow==5.4.1
scipy==1.2.0
Werkzeug==0.14.1
